+++ src/main/java/com/example/shopping/service/AlwaysBuyService.java
@@ -1,49 +1,32 @@
 package com.example.shopping.service;
 
 import java.security.Principal;
-import java.time.LocalDate;
 import java.util.List;
-import java.util.Optional;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 
 import com.example.shopping.entity.AlwaysBuy;
-import com.example.shopping.entity.UserInf;
-import com.example.shopping.repository.AlwaysBuyRepository;
 
 @Service
-public class AlwaysBuyService {
-
-	@Autowired
-	AlwaysBuyRepository alwaysBuyRepository;
-
-	// いつも買うものの一覧を取得
-	public List<AlwaysBuy> get(AlwaysBuy alwaysBuy, Principal principal) {
-		Authentication authentication = (Authentication) principal;
-		UserInf user = (UserInf) authentication.getPrincipal();
-		return alwaysBuyRepository.findAllByUserId(user.getUserId());
-	}
-
-	// いつも買うものの登録
-	public void add(AlwaysBuy alwaysBuy, Principal principal) {
-		AlwaysBuy alwaysBuyEntity = new AlwaysBuy();
-		Authentication authentication = (Authentication) principal;
-		UserInf user = (UserInf) authentication.getPrincipal();
-		alwaysBuyEntity.setUserId(user.getUserId());
-		alwaysBuyEntity.setAlwaysBuyName(alwaysBuy.getAlwaysBuyName());
-		LocalDate createdDate = LocalDate.now();
-		alwaysBuyEntity.setCreatedDate(createdDate);
-		alwaysBuyRepository.saveAndFlush(alwaysBuyEntity);
-	}
-
-	// いつも買うものの削除
-	public void delete(long id) {
-		Optional<AlwaysBuy> findItem = alwaysBuyRepository.findById(id);
-		if (findItem.isPresent()) {
-			findItem.get().setDeleted(true);
-		}
-		alwaysBuyRepository.saveAndFlush(findItem.get());
-	}
+public interface AlwaysBuyService {
+	
+	/**
+	 * いつも買うものの一覧取得
+	 * @param alwaysBuy
+	 * @param principal
+	 * @return
+	 */
+	List <AlwaysBuy> get(AlwaysBuy alwaysBuy, Principal principal);
+	
+	/**
+	 * いつも買うものの登録
+	 * @param alwaysBuy
+	 * @param principal
+	 */
+	void add(AlwaysBuy alwaysBuy, Principal principal);
+	
+	/**
+	 * いつも買うものの削除
+	 * @param id
+	 */
+	void delete(long id);
 }
